v0.0.2/requirements.txt
mysql-connector-python>=8.0.0
psycopg2-binary>=2.9.0
pymongo>=4.0.0
tabulate>=0.9.0
prompt_toolkit>=3.0.0
Pygments>=2.10.0
